+++ CHANGELOG.md
@@ -14,6 +14,9 @@
   - It allows do use asynchronous crypto function like [WebCrypto](https://developer.mozilla.org/en-US/docs/Web/API/Web_Crypto_API)
 - REMOVED many unused plugin hooks because they decreased the performance.
 
+
+- REMOVED support for temporary documents [see here](https://github.com/pubkey/rxdb/pull/3777#issuecomment-1120669088)
+
 <!-- ADD new changes here! -->
 
 <!-- /CHANGELOG NEWEST -->